Feature/top message (#53)
* top_message

* explain how to read cognitive functions

* add top message
add explanation of cognitive functions

---------
