+++ src/index.js
@@ -15,16 +15,6 @@ root.render(
   </React.StrictMode>
 );
 
-// ReactDOM.render(
-//   <React.StrictMode>
-//     <HashRouter>
-//       <App2 />
-//     </HashRouter>
-//   </React.StrictMode>,
-//   document.getElementById('root2')
-// );
-
-
       // If you want to start measuring performance in your app, pass a function
       // to log results (for example: reportWebVitals(console.log))
       // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
